organized vanilla html builder with dups reduction support for non preview builds with module builder class pattern lile react & rn
