+++ packages/builder-web-vanilla/html-css-id-widget/index.ts
@@ -1,0 +1,1 @@
+export * from "./html-css-id-module-builder";
